refactor(cli): split MCP tool catalog
